+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed extra colon in circuit component completion
 - Fixed wrong candidate of function completion
 - Fixed exception when completing circuit expression identifier
+- Added import statement support in Leo
 
 ## [0.2.2]