+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - Fixed extra colon in circuit component completion
 - Fixed wrong candidate of function completion
 - Fixed exception when completing circuit expression identifier
-- Added import statement support in Leo
+- Added import statement grammar support in Leo (Incomplete)
+- Added highlighting for assert calls
 
 ## [0.2.2]